Разделяет код, написанный в main.js на модули
