+++ js/util.js
@@ -1,0 +1,32 @@
+// Функция для получения случайного целого числа
+function getRandomInteger(min, max) {
+  const lower = Math.ceil(Math.min(Math.abs(min), Math.abs(max)));
+  const upper = Math.floor(Math.max(Math.abs(min), Math.abs(max)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+
+  return Math.floor(result);
+}
+
+// Функция-генератор для получения случайных чисел из указанного диапазона
+function createRandomFromRangeGenerator(min, max) {
+  const previousValues = [];
+
+  return function () {
+    let currentValue = getRandomInteger(min, max);
+    if (previousValues.length >= (max - min + 1)) {
+      return null;
+    }
+    while (previousValues.includes(currentValue)) {
+      currentValue = getRandomInteger(min, max);
+    }
+    previousValues.push(currentValue);
+    return currentValue;
+  };
+}
+
+// Функция по поиску случайного элемента в переданном массиве
+const getRandomArrayElement = function (elements) {
+  return elements[getRandomInteger(0, elements.length - 1)];
+};
+
+export { getRandomInteger, createRandomFromRangeGenerator, getRandomArrayElement };
